Move test files do scripts folder
