+++ scripts/test.js
@@ -10,6 +10,7 @@ env.PKG = scope;
 const { DEBUG } = process.env;
 
 const karma = join('node_modules/.bin/karma');
+const karmaConfig = join('scripts/config/karma.conf.js');
 
-exec(`${karma} start --colors` +
+exec(`${karma} start ${karmaConfig} --colors` +
   `${DEBUG ? ' --browsers=Chrome --single-run=false --auto-watch' : ''}`);
